Refactor import path for ContentLayout component
