+++ src/frontend/src/components/layout/content-layout.tsx
